pyproject.toml
[project]
name = "zelador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docker>=7.1.0",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
    "typer>=0.21.1",
]
